configs/requirements.txt
chromedriver-binary-auto==0.1.1
openpyxl==3.0.7
pandas==1.3.1
pip-chill==1.0.1
pywinauto==0.6.8
pyyaml==5.4.1
selenium==3.141.0
